Refactoring: added prediction utilities to prediction.pyx
