+++ python/cuml/cluster/__init__.py
@@ -19,5 +19,5 @@ from cuml.cluster.kmeans import KMeans
 from cuml.cluster.agglomerative import AgglomerativeClustering
 from cuml.cluster.hdbscan import HDBSCAN
 from cuml.cluster.hdbscan import condense_hierarchy
-from cuml.cluster.hdbscan import all_points_membership_vectors
-from cuml.cluster.hdbscan import approximate_predict
+from cuml.cluster.prediction import all_points_membership_vectors
+from cuml.cluster.prediction import approximate_predict
