+++ python/cuml/tests/test_hdbscan.py
@@ -16,7 +16,8 @@
 import pytest
 
 
-from cuml.cluster import HDBSCAN, all_points_membership_vectors,\
+from cuml.cluster import HDBSCAN
+from cuml.cluster.prediction import all_points_membership_vectors,\
     approximate_predict
 from cuml.cluster import condense_hierarchy
 from sklearn.datasets import make_blobs
